ExtractS reusable FeedStoreSpecs helper methods into a shared scope so they can be used by other `FeedStore` implementation tests.
